requirements.txt
appdirs==1.4.4
black==21.5b1
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
plotly==4.14.3
polygon-api-client==0.1.11
psycopg2==2.8.6
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
retrying==1.3.3
six==1.16.0
SQLAlchemy==1.4.14
toml==0.10.2
urllib3==1.26.4
websocket-client==0.59.0
websockets==9.0.1
Werkzeug==1.0.1
